Refactoring of File/Folder Structure -> added UIUtils package
