+++ src/sample/UIUtils/ProgressObserver.java
@@ -1,4 +1,4 @@
-package sample.fileUtils;
+package sample.UIUtils;
 
 public interface ProgressObserver {
     public void update(double d);
